requirements.txt
pyautogui
EasyProcess==1.1
entrypoint2==1.1
jeepney==0.9.0
mss==10.1.0
pillow==11.3.0
py-romanify==1.0.0
pyscreenshot==3.1
